app/requirements.txt
Flask
google-cloud-bigquery
pandas
datetime